Update main_widget.yaml (#29)
* Update main_widget.yaml

Top-Navigation
f7-swiper for mainsection
f7-swiper for floors
f7-swiper for all rooms
f7-swiper for room of selected floor
--> oh-repeaters can be reduced to possibly improve performance.
delete hover color of buttons
add expression to button-class if the objVar isn't set -> Styling like a unselected button

General styling enhancments
remove propertys for style colors until the complete theming isn't intigrated
preparation for custom theming with css classes in stylesheet
css grid for fixing top an bootom navbar
define some css classes for effcient styling
preparation for responsive webdesign in combination with Floors&Rooms -> So the Widgets can be uses on mobile devices, tablets and webbrowsers

Bottom-Navbar
new styling of the navbar (suitable for the concept)
1. Navbar for section0 = Home
2. Navbar for section2 & 3 (Floors & Rooms)


Integration of expresson for set the thingSeletion objVar -> So the Selection is switchable. The button "All" is not needed. All selections can be toggled.
e.G. Select Lights: Only the light items are displayed, Select Lights again: All items are displayed

Icon adjustment of the bottomNavbar to use Iconify. Use different sources is not advised. (svg-container). Preparation to eventually introduce an svg icon set as a local source.

* Fix Swiper responsive breakpoints and more

- topNavbar buttons underline-color in css class filled with variable theme-color
- New paramater for Swiper centerInsufficientSlides: true 
- More breakpoints in the swipers for better display on large screens. The swipers are responsive.
- Swiper arrow color adapted to designSpec

* Adjustment of container

Correct adjustment of the container.
Inclusion of the safe-areas in the container as padding.

* Update openPage.yaml

Necessary adjustment so that the margins of the following block are removed. Below and also above.
The --f7-navbar-height: 0 is not necessary because the paramater hideNavbar: true is already set.